chore: mypy zero errors on 107 source files
Proper type safety fixes across 25 files:

Source (0 mypy errors in 107 files, 1192 tests passing):
- store/connection.py: typed _ConnectionLocal subclass of threading.local
- store/client.py: assert narrowing, typed meta dict
- store/base.py: Self return types, TypeVar generics
- store/state_machine.py: typed transitions ClassVar[list[Transition]]
- agents/_context.py: _ai_instance typed as AI | None (not Any)
- datacube/engine.py: LakehouseProtocol for duck-typed interface
- reactive/computed.py: local var narrowing, Expr | None return type
- agents/_query.py: is_dataclass() guard for fields()
- scheduler/resolve.py: typed Callable local var
- marketdata/server.py: typed TSDBBackend local var
- bridge/store_bridge.py: dict[str, Any] annotation
- streaming/table.py, datacube/server.py: int/bytes casts
- marketdata/client.py, ai/memory.py, ai/team.py: typed returns
- lakehouse/: admin, rls_server, query type fixes
- pyproject.toml: mypy config with python_version 3.12
- tests/__init__.py: fix _sched_fixtures module resolution

+++ agents/_eval/scorers.py
@@ -127,7 +127,7 @@ def score_row_count_preservation(case: AgentEvalCase, artifacts: dict) -> float:
     actual_count = artifacts.get("rows_written", 0)
     if expected_count == 0:
         return 1.0 if actual_count == 0 else 0.0
-    return min(actual_count / expected_count, 1.0)
+    return float(min(actual_count / expected_count, 1.0))
 
 
 def score_star_schema_design(case: AgentEvalCase, artifacts: dict) -> float:
